move shopFor into own helper file
