+++ controllers/itemController.js
@@ -3,20 +3,22 @@ var app = express();
 var Basket = require("../models/basket");
 var Item = require("../models/item");
 var basketController = require("./basketController")
-// var env = require("../env") || ""
 
 var util = require('util');
 OperationHelper = require('../node_modules/apac').OperationHelper;
+// var env = require('../env')
+
+var randomWord = require("./helpers/keyword.js")
 
 
-var randomWord = require("../keyword.js")
-
-
-var opHelper = new OperationHelper({
+var opHelper = new OperationHelper(
+  // env
+  {
   awsId:     process.env.AWS_ID,
   awsSecret: process.env.AWS_SECRET,
   assocId:   process.env.ASSOC_ID
-});
+}
+);
 
 var APICall = function(newBasket, maxPrice, req, res){
   opHelper.execute('ItemSearch', {
@@ -72,8 +74,6 @@ var APICall = function(newBasket, maxPrice, req, res){
 
 
 //
-
-
 function error(response, message){
   response.status(500);
   response.json({error: message})
@@ -102,10 +102,7 @@ var itemController = {
         res.json(basket)
       }
     })
-
-
   }
-
 }
 
 module.exports = itemController
